fix(security): polish/sigma-security-p1p2 — Omega 12 patches, MCP security hardening (#153)
Squash-merges polish/sigma-security-p1p2 (121 commits) onto main.
Conflict resolution: 43 files resolved (HEAD for src, keep-both for tests).

Security patches applied:
- MCP coherence_server: skill validation + fence-escape sanitization
- MCP hookify_server: SurrealQL injection fix
- MCP marimo report: RCE elimination + drop shell=True
- auth.py: formatting cleanup
- DynamicConcurrencyGate singleton reset in conftest.py
- 3 xfail markers for pending SurrealDB/vault integration in test_hookify.py

Also includes Wave 2E/2F type and security annotations across 100+ files.

Entire-Checkpoint: eaa36ebf1a9e

+++ src/cohezion/compound/inflection_detector.py
@@ -243,7 +243,7 @@ class InflectionDetector:
             efficiency_score = max(0.0, min(1.0, cache_hit_rate))
             score *= efficiency_score**0.2
 
-        return round(score, 4)
+        return float(round(score, 4))
 
     def detect_critical(self, result: "ExecutionResult") -> bool:
         """Quick check for critical severity (convenience method).
